feat(user-auth): add util for input validation This inputHandler util removes lot of code duplication. A lot more can be added to this in the future.
