+++ src/resolvers/updateUserResolver.js
@@ -12,6 +12,12 @@ import {
   usernameUpdateAlertEmail,
 } from "../utils/emailForms";
 import { getCurrentDateAndTime } from "../utils/getCurrentDateAndTime";
+import {
+  emailErrorHandler,
+  fullNameErrorHandler,
+  passwordErrorHandler,
+  usernameErrorHandler,
+} from "../utils/inputHandler";
 import { sendEmail } from "../utils/sendEmail";
 
 const updateUserResolver = {
@@ -22,16 +28,8 @@ const updateUserResolver = {
 
         const trimmedNewFullName = newFullName.trim();
 
-        if (!validator.isLength(trimmedNewFullName, { min: 3, max: 50 })) {
-          return {
-            errors: [
-              {
-                field: "fullName",
-                message: "full name must be between 3 and 50 characters long",
-              },
-            ],
-          };
-        }
+        if (fullNameErrorHandler(trimmedNewFullName).errors)
+          return fullNameErrorHandler(trimmedNewFullName);
 
         if (trimmedNewFullName === fullName) {
           return {
@@ -62,17 +60,8 @@ const updateUserResolver = {
       try {
         const { username } = await User.findById(req.session.userId);
 
-        if (!validator.isAlphanumeric(newUsername, "en-US", { ignore: "_-" })) {
-          return {
-            errors: [
-              {
-                field: "username",
-                message:
-                  "username can only contain letters, numbers, underscores (_) and dashes (-)",
-              },
-            ],
-          };
-        }
+        if (usernameErrorHandler(newUsername).errors)
+          return usernameErrorHandler(newUsername);
 
         if (username === newUsername) {
           return {
@@ -160,26 +149,8 @@ const updateUserResolver = {
 
     updateForgotPassword: async (_, { token, newPassword }, { redis, req }) => {
       try {
-        if (
-          !validator.isStrongPassword(newPassword, {
-            minLength: 8,
-            minLowercase: 1,
-            minUppercase: 1,
-            minNumbers: 1,
-            minSymbols: 1,
-            returnScore: false,
-          })
-        ) {
-          return {
-            errors: [
-              {
-                field: "password",
-                message:
-                  "password must be at least 8 characters long and contain at least 1 lowercase letter, 1 uppercase letter, 1 number and 1 symbol",
-              },
-            ],
-          };
-        }
+        if (passwordErrorHandler(newPassword).errors)
+          return passwordErrorHandler(newPassword);
 
         const key = FORGET_PASSWORD_PREFIX + token;
         const userId = await redis.get(key);
@@ -241,26 +212,8 @@ const updateUserResolver = {
           };
         }
 
-        if (
-          !validator.isStrongPassword(newPassword, {
-            minLength: 8,
-            minLowercase: 1,
-            minUppercase: 1,
-            minNumbers: 1,
-            minSymbols: 1,
-            returnScore: false,
-          })
-        ) {
-          return {
-            errors: [
-              {
-                field: "password",
-                message:
-                  "password must be at least 8 characters long and contain at least 1 lowercase letter, 1 uppercase letter, 1 number and 1 symbol",
-              },
-            ],
-          };
-        }
+        if (passwordErrorHandler(newPassword).errors)
+          return passwordErrorHandler(newPassword);
 
         const currentUser = await User.findByIdAndUpdate(
           req.session.userId,
@@ -277,6 +230,90 @@ const updateUserResolver = {
         );
 
         return { user: updatedUser };
+      } catch (error) {
+        return error;
+      }
+    },
+
+    sendNewEmailAddressVerificationEmail: async (
+      _,
+      { newEmailAddress, password },
+      { redis, req }
+    ) => {
+      try {
+        const currentUser = await User.findById(req.session.userId);
+
+        // check if email address is valid
+        if (emailErrorHandler(newEmailAddress).errors)
+          return emailErrorHandler(newEmailAddress);
+
+        // check if the new email address is the same as the old one
+        if (newEmailAddress === currentUser.email) {
+          return {
+            errors: [
+              {
+                field: "newEmailAddress",
+                message: "new email address cannot be the same as the old one",
+              },
+            ],
+          };
+        }
+
+        // check if the new email address is already taken
+        const isEmailTaken = await User.findOne({ email: newEmailAddress });
+        if (isEmailTaken) {
+          return {
+            errors: [
+              {
+                field: "newEmailAddress",
+                message: "email address already taken",
+              },
+            ],
+          };
+        }
+
+        // verify password
+        const isOldPasswordCorrect = await bcrypt.compare(
+          password,
+          currentUser.password
+        );
+
+        if (!isOldPasswordCorrect) {
+          return {
+            errors: [
+              {
+                field: "password",
+                message: "old password does not match",
+              },
+            ],
+          };
+        }
+
+        const token = v4();
+        await redis.set(
+          VERIFY_NEW_EMAIL_ADDRESS_PREFIX + token,
+          currentUser.id,
+          "ex",
+          1000 * 60 * 60 * 24 * 1 // 1 day
+        );
+
+        await redis.set(
+          NEW_EMAIL_ADDRESS_PREFIX + token,
+          newEmailAddress,
+          "ex",
+          1000 * 60 * 60 * 24 * 1 // 1 day
+        );
+
+        // send verification email to new email address
+        const emailResult = await sendEmail(
+          newEmailAddress,
+          "Confirm New Email Address",
+          verifyNewEmailAddressEmail(
+            `${process.env.CORS_ORIGIN}/new-email-verification/${token}`
+          )
+        );
+
+        return { user: currentUser };
       } catch (error) {
         return error;
       }
